+++ packages/eslint-plugin/lib/configs/rules/possible-errors.js
@@ -1,3 +1,7 @@
 // https://eslint.org/docs/rules/#possible-errors
 
-module.exports = {};
+module.exports = {
+	// Disallow calling some `Object.prototype` methods directly on objects
+	// https://eslint.org/docs/rules/no-prototype-builtins
+	'no-prototype-builtins': 'error',
+};